% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoairline.R
\name{geoairline}
\alias{geoairline}
\title{geoairline}
\usage{
geoairline(city1, country1, city2, country2)
}
\arguments{
\item{city1}{english name of the first city}

\item{country1}{english name of the country of the first city}

\item{city2}{english name of the second city}

\item{country2}{english name of the country of the second city}
}
\value{
object of class geoairline - list of two elements: cities_sf and LINESTRING geometry between the cities
}
\description{
geoairline
}
\examples{
berlin2tokyo <- geoairline("Berlin", "Germany", "Tokyo", "Japan")
}
