% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoairline_addcity.R
\name{geoairline_addcity}
\alias{geoairline_addcity}
\title{geoairline_addcity}
\usage{
geoairline_addcity(geoairline, city, country)
}
\arguments{
\item{geoairline}{object of class geoairline where to add the new city to}

\item{city}{city name in english of the new city}

\item{country}{country name in english of the new city}
}
\value{
geoairline
}
\description{
This function adds another city to an object of class geoairline
}
\examples{
b2t <- geoairline("Berlin", "Germany", "Tokyo", "Japan")
b2t2r <- geoairline_addcity(b2t, "Rio de Janeiro", "Brazil")
}
