% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolinecompare.R
\name{geolinecompare}
\alias{geolinecompare}
\title{geolinecompare}
\usage{
geolinecompare(sfline_wgs84)
}
\arguments{
\item{sfline_wgs84}{a LINESTRING simple feature in WGS84}
}
\value{
object of class geolinecompare
}
\description{
This function takes a LINESTRING simple feature in WGS84 and creates its copy in an equirectangular projection as if the ellipsoidal coordinates are treated as metrics and not as spherical. It returns a list as s3 class geoline_compare
}
\examples{
random_matrix <- matrix(c(runif(30, -90, 90), runif(30, -180, 180)), nrow=10, ncol=2)
random_line <- sf::st_linestring(random_matrix)
random_line <- sf::st_sfc(random_line, crs = 4326)
geolinecompare <- geolinecompare(random_line)

}
