% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolinecompare_diff.R
\name{geolinecompare_diff}
\alias{geolinecompare_diff}
\title{geolinecompare_diff}
\usage{
geolinecompare_diff(x)
}
\arguments{
\item{x}{object of class geolinecompare}
}
\value{
length in meter
}
\description{
This function calculates the length difference between a sf LINESTRING on WGS84 and the equirectangular projection
}
\examples{
random_matrix <- matrix(c(runif(30, -90, 90), runif(30, -180, 180)), nrow=10, ncol=2)
random_line <- sf::st_linestring(random_matrix)
random_line <- sf::st_sfc(random_line, crs = 4326)
random_line_comp <- geolinecompare(random_line)
geolinecompare_diff(random_line_comp)

}
