% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_globe.R
\name{ps_globe}
\alias{ps_globe}
\title{Create the globe}
\usage{
ps_globe(show_poles = TRUE, alpha = 1, ...)
}
\arguments{
\item{show_poles}{logical if the poles should be added for orientation}

\item{alpha}{set transparency for the globe
0 = transparent
1 = non-transparent}

\item{...}{more parameters of rgl::material3d()}
}
\description{
Plots the basic globe in a rgl window device
}
\examples{
ps_globe(show_poles=TRUE)
}
