% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_lines.R
\name{ps_lines}
\alias{ps_lines}
\title{Plot simple feature lines on a sphere}
\usage{
ps_lines(sf_lines, color = "black", alpha = 1, ...)
}
\arguments{
\item{sf_lines}{simple feature linestrings or mutli-linestrings}

\item{color}{set color of the point(s)}

\item{alpha}{set transparency of the point(s)
0 = transparent
1 = non-transparent}

\item{...}{more parameters of rgl::material3d()}
}
\description{
Plot simple feature lines on a sphere. Use ps_globe() first to plot the lines on the globe
}
\examples{
random_matrix <- matrix(c(runif(30, -90, 90), runif(30, -180, 180)), nrow=10, ncol=2)
random_line <- sf::st_linestring(random_matrix)
random_line <- sf::st_sfc(random_line, crs = 4326)
ps_globe()
ps_lines(random_line, color = 'green')
}
