% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_points.R
\name{ps_points}
\alias{ps_points}
\title{Plot simple feature points on a sphere}
\usage{
ps_points(sf_points, color = "black", alpha = 1, ...)
}
\arguments{
\item{sf_points}{simple feature point or multi-points}

\item{color}{set color of the point(s)}

\item{alpha}{set transparency of the point(s)
0 = transparent
1 = non-transparent}

\item{...}{more parameters of rgl::material3d()}
}
\description{
Plot simple feature points on a sphere. Use ps_globe() first to plot the points on the globe
}
\examples{
random_matrix <- matrix(c(runif(30, -90, 90), runif(30, -180, 180)), nrow=10, ncol=2)
random_points <- sf::st_multipoint(random_matrix)
random_points <- sf::st_sfc(random_points, crs = 4326)
ps_globe()
ps_points(random_points, color='red', alpha=0.8)
}
