% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_polygons.R
\name{ps_polygons}
\alias{ps_polygons}
\title{Plot simple feature polygons on a sphere}
\usage{
ps_polygons(sf_polygons, color = "black", alpha = 1, ...)
}
\arguments{
\item{sf_polygons}{simple feature polygons or mutli-polygons}

\item{color}{set color of the point(s)}

\item{alpha}{set transparency of the point(s)
0 = transparent
1 = non-transparent}

\item{...}{more parameters of rgl::material3d()}
}
\description{
Plot simple feature polygons on a sphere. Use ps_globe() first to plot the polygons on the globe
}
\examples{
poly <- sf::st_polygon(list(rbind(c(0,23), c(3,21), c(15,32), c(26,12),
c(25,-33), c(42,-40), c(32,-54), c(10,-23), c(0,-23), c(0,23))))
poly <- sf::st_sfc(poly, crs = 4326)
ps_globe(alpha=0.4)
ps_polygons(poly, color = 'darkgreen')
}
