% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world.cities.R
\docType{data}
\name{world.cities}
\alias{world.cities}
\title{World cities data}
\format{
A data frame with 43645 rows and 6 variables (including latitude and longitude):
\describe{
\item{name}{name, english}
\item{country.etc}{country name, english}
\item{pop}{population count}
\item{lat}{latitude in degree}
\item{long}{longitude in degree}
\item{capital}{binary}
}
}
\source{
\url{https://www.rdocumentation.org/packages/maps/versions/3.3.0/topics/world.cities}
}
\usage{
world.cities
}
\description{
A dataset containing world cities of population greater than about 40,000. Also included are capital cities of any population size, and many smaller towns.
}
\keyword{datasets}
